Add: roi selector
