+++ modules/tracking/include/opencv2/tracking/tracker.hpp
@@ -1322,6 +1322,41 @@ class CV_EXPORTS_W MultiTracker
   String defaultAlgorithm;
 };
 
+class ROISelector {
+public:
+  Rect2d select(Mat img, bool fromCenter = true);
+  Rect2d select(const std::string& windowName, Mat img, bool showCrossair = true, bool fromCenter = true);
+  void select(const std::string& windowName, Mat img, std::vector<Rect2d> & boundingBox, bool fromCenter = true);
+
+  struct handlerT{
+    // basic parameters
+    bool isDrawing;
+    Rect2d box;
+    Mat image;
+
+    // parameters for drawing from the center
+    bool drawFromCenter;
+    Point2f center;
+
+    // initializer list
+    handlerT(): isDrawing(false), drawFromCenter(true) {};
+  }selectorParams;
+
+  // to store the tracked objects
+  std::vector<handlerT> objects;
+
+private:
+  static void mouseHandler(int event, int x, int y, int flags, void *param);
+  void opencv_mouse_callback( int event, int x, int y, int , void *param );
+
+  // save the keypressed characted
+  int key;
+};
+
+Rect2d CV_EXPORTS_W selectROI(Mat img, bool fromCenter = true);
+Rect2d CV_EXPORTS_W selectROI(const std::string& windowName, Mat img, bool showCrossair = true, bool fromCenter = true);
+void CV_EXPORTS_W selectROI(const std::string& windowName, Mat img, std::vector<Rect2d> & boundingBox, bool fromCenter = true);
+
 } /* namespace cv */
 
 #endif
